(BSR)[PRO] refactor: createImageFile to testHelpers
