+++ pro/src/components/pages/Offers/Offer/Thumbnail/_constants.js
@@ -2,7 +2,6 @@ export const IMPORT_TAB_ID = 'import'
 export const IMAGE_TYPE = ['image/png', 'image/jpeg']
 export const MAX_IMAGE_SIZE = 10_000_000
 export const MIN_IMAGE_WIDTH = 400
-export const MIN_IMAGE_HEIGHT = 400
 export const CANVAS_HEIGHT = 384
 export const CANVAS_WIDTH = (CANVAS_HEIGHT * 6) / 9
 export const CROP_BORDER_HEIGHT = 50
